version 3.2.2
1). Adding some lines of javascript in script.js file so to fix a bug.
The bug was basically =>Earlier a hovering effect was added via css in two buttons (like and hire us) in the Hire us section. Which was not properly working as the buttons did not changed the color at the same time while hoveing. So removed those css lines and solved it via javascript.
2). Also added some more lines so as to meet the desired input in forms section for both the categories (Employer, Job Seeker).